simplify test runner (now breaks some tests because empty files are generated instead of no file at all in case of errors)

+++ src/Program.cs
@@ -3,6 +3,7 @@ using System.Configuration;
 using System.Data.Common;
 using System.Data.SqlClient;
 using System.IO;
+using System.Linq;
 using System.Text.RegularExpressions;
 using DatabaseSchemaReader;
 using TCode.r2rml4net.Mapping;
@@ -10,8 +11,7 @@ using TCode.r2rml4net.Mapping.DirectMapping;
 using TCode.r2rml4net.RDB.DatabaseSchemaReader;
 using TCode.r2rml4net.TriplesGeneration;
 using VDS.RDF;
-using VDS.RDF.Parsing.Handlers;
-using VDS.RDF.Writing.Formatting;
+using VDS.RDF.Writing;
 using System.Data;
 
 namespace TCode.r2rml4net.TestCasesRunner
@@ -19,11 +19,11 @@ namespace TCode.r2rml4net.TestCasesRunner
     class Program
     {
         static readonly IColumnTypeMapper ColumnTypeMapper = new MSSQLServerColumTypeMapper();
-        private static StreamWriter _r2rmlLog, _directLog;
+        private static StreamWriter _r2RMLLog, _directLog;
 
-        static void Main(string[] args)
+        static void Main()
         {
-            _r2rmlLog = new StreamWriter("r2rml.log");
+            _r2RMLLog = new StreamWriter("r2rml.log");
             _directLog = new StreamWriter("direct.log");
 
             string masterConnection = ConfigurationManager.ConnectionStrings["SqlServer2008Master"].ConnectionString;
@@ -34,34 +34,36 @@ namespace TCode.r2rml4net.TestCasesRunner
                 Console.Out.WriteLine("Test case {0}: ", testCase);
 
                 var dbProviderFactory = DbProviderFactories.GetFactory(ConfigurationManager.ConnectionStrings["SqlServer2008Master"].ProviderName);
-                using (DbConnection connection = dbProviderFactory.CreateConnection())
+                using (IDbConnection connection = dbProviderFactory.CreateConnection())
                 {
-                    connection.ConnectionString = masterConnection;
-                    connection.Open();
-                    try
+                    if (connection != null)
                     {
-                        ExecuteCommand(connection, "if db_id('r2rml4net_tests') is not null DROP DATABASE r2rml4net_tests");
-                        ExecuteCommand(connection, "CREATE DATABASE r2rml4net_tests");
+                        connection.ConnectionString = masterConnection;
+                        connection.Open();
+                        try
+                        {
+                            ExecuteCommand(connection, "if db_id('r2rml4net_tests') is not null DROP DATABASE r2rml4net_tests");
+                            ExecuteCommand(connection, "CREATE DATABASE r2rml4net_tests");
 
-                        connection.ChangeDatabase("r2rml4net_tests");
+                            connection.ChangeDatabase("r2rml4net_tests");
 
-                        ExecuteTests(connection, testCase);
-                    }
-                    catch (SqlException ex)
-                    {
-                        Console.WriteLine("FAIL");
-                        Console.WriteLine(ex.Message);
-                    }
-                    finally
-                    {
-                        connection.ChangeDatabase("master");
-                        ExecuteCommand(connection, "ALTER DATABASE r2rml4net_tests SET SINGLE_USER WITH ROLLBACK IMMEDIATE");
-                        ExecuteCommand(connection, "ALTER DATABASE r2rml4net_tests SET MULTI_USER");
+                            ExecuteTests(connection, testCase);
+                        }
+                        catch (SqlException ex)
+                        {
+                            Console.WriteLine("FAIL");
+                            Console.WriteLine(ex.Message);
+                        }
+                        finally
+                        {
+                            connection.ChangeDatabase("master");
+                            ExecuteCommand(connection, "ALTER DATABASE r2rml4net_tests SET SINGLE_USER WITH ROLLBACK IMMEDIATE");
+                            ExecuteCommand(connection, "ALTER DATABASE r2rml4net_tests SET MULTI_USER");
+                        }
                     }
                 }
 
                 Console.WriteLine();
-
             }
         }
 
@@ -74,7 +76,7 @@ namespace TCode.r2rml4net.TestCasesRunner
             }
         }
 
-        private static void ExecuteTests(DbConnection connection, string testCaseDirectory)
+        private static void ExecuteTests(IDbConnection connection, string testCaseDirectory)
         {
             string sqlScript = Path.Combine(testCaseDirectory, "create.sql");
             string directOutput = Path.Combine(testCaseDirectory, "directGraph-r2rml4net.ttl");
@@ -94,111 +96,51 @@ namespace TCode.r2rml4net.TestCasesRunner
             }
         }
 
-        private static void ExecuteR2RMLTest(DbConnection connection, string inputMappingPath, string outputDatasetPath)
+        private static void ExecuteR2RMLTest(IDbConnection connection, string inputMappingPath, string outputDatasetPath)
         {
             Console.Out.Write("R2RML: ");
 
-            var countHandler = new CountHandler();
-
-            _r2rmlLog.WriteLine();
-            _r2rmlLog.WriteLine(inputMappingPath);
-            ErrorCounterLog log = new ErrorCounterLog(_r2rmlLog);
-            try
-            {
-                using (var directStream = File.Create(outputDatasetPath))
-                {
-                    using (var streamWriter = new StreamWriter(directStream))
-                    {
-                        IRdfHandler writer = new ChainedHandler(new IRdfHandler[]
-                                {
-                                    countHandler, 
-                                    new WriteThroughHandler(new NQuadsFormatter(), streamWriter)
-                                });
-                        IRDFTermGenerator termGen = new RDFTermGenerator
-                        {
-                            Log = log
-                        };
-                        IR2RMLProcessor procesor = new W3CR2RMLProcessor(connection, termGen)
-                            {
-                                Log = log
-                            };
-                        var mappings = R2RMLLoader.Load(File.OpenRead(inputMappingPath));
-                        procesor.GenerateTriples(mappings, writer);
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                Console.Out.WriteLine("FAIL! {0}", ex.Message);
-                return;
-            }
-
-            ReportResult(outputDatasetPath, countHandler, log);
+            _r2RMLLog.WriteLine();
+            _r2RMLLog.WriteLine(inputMappingPath);
+            var mappings = new Func<IR2RML>(() => R2RMLLoader.Load(File.OpenRead(inputMappingPath)));
+            GenerateTriples(mappings, outputDatasetPath, connection, new RDFTermGenerator());
         }
 
-        private static void ExecuteDirectMappingTest(DbConnection connection, string directMappingOutputPath)
+        private static void ExecuteDirectMappingTest(IDbConnection connection, string directMappingOutputPath)
         {
             Console.Out.Write("DIRECT: ");
 
-            var countHandler = new CountHandler();
-
             _directLog.WriteLine();
             _directLog.WriteLine(directMappingOutputPath);
-            var log = new ErrorCounterLog(_directLog);
+            var connectionString = ConfigurationManager.ConnectionStrings["SqlServer2008Test"];
+            using (var databaseReader = new DatabaseReader(connectionString.ConnectionString, connectionString.ProviderName))
+            {
+                var r2RMLConfiguration = new R2RMLConfiguration(new Uri("http://example.com/base/"));
+                var metadataProvider = new DatabaseSchemaAdapter(databaseReader, ColumnTypeMapper);
+                var mappingGenerator = new R2RMLMappingGenerator(metadataProvider, r2RMLConfiguration);
+
+                var mappings = new Func<IR2RML>(mappingGenerator.GenerateMappings);
+                GenerateTriples(mappings, directMappingOutputPath, connection, new RDFTermGenerator(true));
+            }
+        }
+
+        private static void GenerateTriples(Func<IR2RML> createMappings, string outPath, IDbConnection connection, IRDFTermGenerator termGen)
+        {
+            File.Delete(outPath);
+            ITripleStore store;
             try
             {
-                using (var directStream = File.Create(directMappingOutputPath))
-                {
-                    using (var streamWriter = new StreamWriter(directStream))
-                    {
-                        var connectionString = ConfigurationManager.ConnectionStrings["SqlServer2008Test"];
-                        using (var databaseReader = new DatabaseReader(connectionString.ConnectionString, connectionString.ProviderName))
-                        {
-                            var r2RMLConfiguration = new R2RMLConfiguration(new Uri("http://example.com/base/"));
-                            var metadataProvider = new DatabaseSchemaAdapter(databaseReader, ColumnTypeMapper);
-                            var mappingGenerator = new R2RMLMappingGenerator(metadataProvider, r2RMLConfiguration);
-
-                            var mappings = mappingGenerator.GenerateMappings();
-
-                            IRdfHandler writer = new ChainedHandler(new IRdfHandler[]
-                                {
-                                    countHandler, 
-                                    new WriteThroughHandler(new TurtleW3CFormatter(), streamWriter)
-                                });
-                            IRDFTermGenerator termGen = new RDFTermGenerator(true)
-                                {
-                                    Log = log
-                                };
-                            IR2RMLProcessor procesor = new W3CR2RMLProcessor(connection, termGen)
-                                {
-                                    Log = log
-                                };
-                            procesor.GenerateTriples(mappings, writer);
-                        }
-                    }
-                }
+                IR2RMLProcessor processor = new W3CR2RMLProcessor(connection, termGen);
+                store = processor.GenerateTriples(createMappings());
             }
             catch (Exception ex)
             {
                 Console.Out.WriteLine("FAIL! {0}", ex.Message);
-                File.Delete(directMappingOutputPath);
                 return;
             }
 
-            ReportResult(directMappingOutputPath, countHandler, log);
-        }
-
-        private static void ReportResult(string outputPath, CountHandler countHandler, ErrorCounterLog log)
-        {
-            if (log.ErrorsCount == 0)
-            {
-                Console.Out.WriteLine("SUCCESS! {0} triples generated", countHandler.Count);
-            }
-            else
-            {
-                Console.Out.WriteLine("SUCCESS! No output file produced");
-                File.Delete(outputPath);
-            }
+            store.SaveToFile(outPath, new NQuadsWriter());
+            Console.Out.WriteLine("SUCCESS! {0} triples generated", store.Triples.Count());
         }
     }
 }
